refactor(wsman): added new wsman message creator
